requirements.txt
asgiref==3.3.1
Django==3.1.7
django-bootstrap-form==3.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
xlwt==1.3.0
